[update] event_loop -> events & fix all references/imports to the Event.
